+++ test/test_examples.py
@@ -54,13 +54,13 @@ howto_list = {
     "bf_mt_002": "mlpro.bf.examples.howto_bf_mt_002_tasks_and_workflows",
     "bf_math_001": "mlpro.bf.examples.howto_bf_math_001_spaces_and_elements",
     "bf_math_010": "mlpro.bf.examples.howto_bf_math_010_normalizers",
+    "bf_math_020": "mlpro.bf.examples.howto_bf_math_020_unit_converter.py",
     "bf_ml_001": "mlpro.bf.examples.howto_bf_ml_001_hyperparameters",
     # "bf_streams_051": "mlpro.bf.examples.howto_bf_streams_051_accessing_data_from_openml",
     "bf_streams_052": "mlpro.bf.examples.howto_bf_streams_052_accessing_data_from_scikitlearn",
     "bf_streams_053": "mlpro.bf.examples.howto_bf_streams_053_accessing_data_from_river",
     "bf_systems_001": "mlpro.bf.examples.howto_bf_systems_001_systems_controllers_actuators_sensors",
     "bf_systems_002": "mlpro.bf.examples.howto_bf_systems_002_set_up_transfer_functions",
-    "bf_systems_003": "mlpro.bf.examples.howto_bf_systems_003_unit_converter",
 
 # Reinforcement Learning:
     "rl_001": "mlpro.rl.examples.howto_rl_001_reward",
